+++ gips/test/sys/util.py
